Fixes, improvements
