+++ src/main/java/at/hannibal2/skyhanni/config/features/chat/FilterTypesConfig.java
@@ -14,6 +14,11 @@ public class FilterTypesConfig {
     public PowderMiningFilterConfig powderMiningFilter = new PowderMiningFilterConfig();
 
     @Expose
+    @ConfigOption(name = "Stash Messages", desc = "")
+    @Accordion
+    public StashConfig stashMessages = new StashConfig();
+
+    @Expose
     @ConfigOption(name = "Hypixel Lobbies", desc = "Hide announcements in Hypixel lobbies " +
         "(player joins, loot boxes, prototype lobby messages, radiating generosity, Hypixel tournaments)")
     @ConfigEditorBoolean
