+++ src/main/java/at/hannibal2/skyhanni/utils/HypixelCommands.kt
@@ -162,6 +162,10 @@ object HypixelCommands {
         send("pickupstash")
     }
 
+    fun viewStash(type: String) {
+        send("viewstash $type")
+    }
+
     private fun send(command: String) {
         @Suppress("DEPRECATION")
         // TODO rename function
